toggleTurn/checkForWin/resetBoard
Finalized the JS logic for these features
(toggleTurn/checkForWin/resetBoard).

The board has an ID set for each cell. With these IDs, and event handler
s, I was able to listen for clicks on each cell. On a click, the
ID is displayed. Once I was able to display the cell clicked in the
console, I was able to create some logic to toggle between the X turn
and O's turn. X is the default starting turn. I have a function that
checks an array of 0's (one for each cell). If this is true,
and the coordinate that was clicked has an index of 0, then the
gamestate is updated to 1 for X's and 2 for O's, X is updated to false
(which switches the player)
and then it checks if any neighboring indexes that
match (checkForWin) the
current gamestate #. This process is repeated for O's until a
turnCounter reaches 8 turns, if none of the conditions have been met,
and the turnCounter reaches it's maximum amount of turns then an
alert pops up telling players that there has been a draw.

next steps: replace alerts to modals, create a way to display and hold
wins or rounds played. But only after setting up AJAX to get
my sign-in, sign-up, and sign-out buttons and communicating with the
API.