fixed patch request
the only way it would work is if I put the patch request in the same
file as toggleTurn. Previously it was in it's own file but it kept
throwing errors, likely when requiring or exporting, and would not
recognize updateServer as a function. This was resolved but putting it
in the same module that held toggleTurn the function that calls it.