show/hide and gameplay optimized
The reset button is only available during game play and when clicked, it
clears the board, and sets gameOver to true while directing players to
start a new game.

Moved the gameStats button, game stats notifications, and change pass
notifications to the left column. These features are only available
when clicked and hide after game play begins or a user signs out.

Tested the game and buttons, as well as user interaction flow, in every
possible scenario I could think of. Everything seems to be working
properly.

If possible I will be giving the link to a friend for further testing,
once I get their feedback, I will submit this through email tonight to
meet the resubmission deadline.